refactored code to remove one level of component children. Reduced complexity.
