+++ src/components/Timeline/index.ts
@@ -1,2 +1,1 @@
 export * from './Timeline2'
-export * from './Lane'
